+++ app/src/main/java/com/github/droidworksstudio/mlauncher/ui/settings/GesturesFragment.kt
@@ -1,0 +1,149 @@
+package com.github.droidworksstudio.mlauncher.ui.settings
+
+import android.app.admin.DevicePolicyManager
+import android.content.ComponentName
+import android.content.Context
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.TextUnit
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.isSpecified
+import androidx.compose.ui.unit.sp
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.fragment.findNavController
+import com.github.droidworksstudio.mlauncher.MainViewModel
+import com.github.droidworksstudio.mlauncher.R
+import com.github.droidworksstudio.mlauncher.data.Constants
+import com.github.droidworksstudio.mlauncher.data.Constants.Theme.Dark
+import com.github.droidworksstudio.mlauncher.data.Constants.Theme.Light
+import com.github.droidworksstudio.mlauncher.data.Constants.Theme.System
+import com.github.droidworksstudio.mlauncher.data.Prefs
+import com.github.droidworksstudio.mlauncher.databinding.FragmentSettingsBinding
+import com.github.droidworksstudio.mlauncher.helper.getHexForOpacity
+import com.github.droidworksstudio.mlauncher.helper.isSystemInDarkMode
+import com.github.droidworksstudio.mlauncher.helper.setThemeMode
+import com.github.droidworksstudio.mlauncher.listener.DeviceAdmin
+import com.github.droidworksstudio.mlauncher.style.SettingsTheme
+import com.github.droidworksstudio.mlauncher.ui.compose.SettingsComposable.PageHeader
+
+class GesturesFragment : Fragment() {
+
+    private lateinit var prefs: Prefs
+    private lateinit var viewModel: MainViewModel
+    private lateinit var deviceManager: DevicePolicyManager
+    private lateinit var componentName: ComponentName
+
+    private var _binding: FragmentSettingsBinding? = null
+    private val binding get() = _binding!!
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentSettingsBinding.inflate(inflater, container, false)
+        prefs = Prefs(requireContext())
+        val backgroundColor = getHexForOpacity(prefs)
+        binding.scrollView.setBackgroundColor(backgroundColor)
+        return binding.root
+    }
+
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        resetThemeColors()
+    }
+
+    @Composable
+    private fun Settings(fontSize: TextUnit = TextUnit.Unspecified) {
+        val fs = remember { mutableStateOf(fontSize) }
+        Constants.updateMaxHomePages(requireContext())
+
+        val titleFontSize = if (fs.value.isSpecified) {
+            (fs.value.value * 1.5).sp
+        } else fs.value
+
+        val descriptionFontSize = if (fs.value.isSpecified) {
+            (fs.value.value * 1.2).sp
+        } else fs.value
+
+
+        Column {
+            PageHeader(
+                iconRes = R.drawable.ic_back,
+                title = stringResource(R.string.gestures_settings_title),
+                onClick = { goBackToLastFragment() }
+            )
+
+            Spacer(
+                modifier = Modifier
+                    .height(16.dp)
+            )
+        }
+    }
+
+    private fun resetThemeColors() {
+        binding.settingsView.setContent {
+
+            val isDark = when (prefs.appTheme) {
+                Light -> false
+                Dark -> true
+                System -> isSystemInDarkMode(requireContext())
+            }
+
+            setThemeMode(requireContext(), isDark, binding.scrollView)
+            val settingsSize = (prefs.settingsSize - 3)
+
+            SettingsTheme(isDark) {
+                Settings(settingsSize.sp)
+            }
+        }
+    }
+
+    private fun goBackToLastFragment() {
+        findNavController().popBackStack()
+    }
+
+    private fun dismissDialogs() {
+//        backupRestoreDialog?.dismiss()
+    }
+
+    @Deprecated("Deprecated in Java")
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        @Suppress("DEPRECATION")
+        super.onActivityCreated(savedInstanceState)
+        prefs = Prefs(requireContext())
+        viewModel = activity?.run {
+            ViewModelProvider(this)[MainViewModel::class.java]
+        } ?: throw Exception("Invalid Activity")
+
+        viewModel.ismlauncherDefault()
+
+        deviceManager =
+            context?.getSystemService(Context.DEVICE_POLICY_SERVICE) as DevicePolicyManager
+        componentName = ComponentName(requireContext(), DeviceAdmin::class.java)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    override fun onStop() {
+        super.onStop()
+        dismissDialogs()
+    }
+}